Move SDKConnectors SDKConnection class into seperate file
The next will be to merge SDKConnector into the ScreenPlayManager.

We now have the first version of bi directional messages. This is for updating values like
saving the current Widget position. We also save a SDKConnection shared reference inside
our ScreenPlayWallpaper or ScreenPlayWidget instance. So we now have all logic in these classes.