Add ScreenPlayCoreWindowsExtrasLib
We hackily used the class in the main
ScreenPlayApp. Because it is also used in
Godot, we make it an separate lib, so we
do not include Qt

Add monitor name. This is not that easy, because
Windows does essentially not give our any useful
name information. The way it works is to parse
EDID data. See https://github.com/bsdhw/EDID/tree/master
for future reference. At least we cover some hardware
vendors for now.